02_anthropic/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "anthropic-api-labs"
version = "0.1.0"
description = "Educational Anthropic API labs with scripts and notebooks."
readme = "README.md"
requires-python = ">=3.13.12"
dependencies = [
    "anthropic>=0.49.0",
    "python-dotenv>=1.0.1",
    "voyageai>=0.3.0",
    "pydantic>=2.8.0",
]

[project.optional-dependencies]
notebooks = [
    "ipykernel>=6.29.0",
    "jupyterlab>=4.2.0",
]

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
